+++ src/app/components/team/TodoList.tsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import { useQuery } from '@tanstack/react-query';
-import { PieChart, Pie } from 'recharts';
-import getTaskList, { Task } from '@/app/lib/group/getTaskList';
-import Link from 'next/link';
-import IconTaskDone from '../icons/IconTaskDone';
-import DropdownMenu from './DropdownMenu';
+import getTaskList from '@/app/lib/group/getTaskList';
+import TodoListItem from '@/app/components/team/TodoListItem';
 
 interface TodoListProps {
   groupId: number;
@@ -45,15 +42,25 @@ export default function TodoList({ groupId, taskLists = [] }: TodoListProps) {
   });
 
   if (isLoading) {
-    return <div>로딩 중...</div>;
+    return <div className="mx-auto my-6 max-w-[75rem]">로딩 중...</div>;
   }
 
   if (isError || !data) {
-    return <div>데이터를 불러오는 중 오류가 발생했습니다.</div>;
+    return (
+      <div className="mx-auto my-6 max-w-[75rem]">
+        데이터를 불러오는 중 오류가 발생했습니다.
+      </div>
+    );
   }
 
   if (taskLists.length === 0) {
-    return <div>할 일이 없습니다.</div>;
+    return (
+      <div className="mx-auto my-6 flex max-w-[75rem] items-center justify-center">
+        <div className="text-md font-medium text-text-default">
+          아직 할 일 목록이 없습니다.
+        </div>
+      </div>
+    );
   }
 
   return (
@@ -66,91 +73,16 @@ export default function TodoList({ groupId, taskLists = [] }: TodoListProps) {
           + 새로운 목록 추가하기
         </button>
       </div>
-      {taskLists.map((taskList, index) => {
-        const taskListData = data[index];
 
-        if (!taskListData) {
-          return (
-            <div
-              key={taskList.id}
-              className="relative mt-4 flex h-10 w-full items-center justify-between rounded-xl bg-background-secondary pl-6 pr-2"
-            >
-              <div
-                className={`absolute left-0 h-10 w-3 rounded-l-xl ${backgroundColors[index % backgroundColors.length]}`}
-              />
-              <span className="text-white">{taskList.name}</span>
-              <span className="text-sm font-bold text-gray-600">
-                로딩 중...
-              </span>
-              <DropdownMenu iconType="task" />
-            </div>
-          );
-        }
-
-        const completedItems = taskListData.tasks.filter(
-          (task: Task) => task.doneAt !== null,
-        ).length;
-        const totalTasks = taskListData.tasks.length;
-        const completionPercentage = (completedItems / totalTasks) * 100;
-
-        return (
-          <div
-            key={taskList.id}
-            className="relative mt-4 flex h-10 w-full items-center justify-between rounded-xl bg-background-secondary pl-6 pr-4"
-          >
-            <div
-              className={`absolute left-0 h-10 w-3 rounded-l-xl ${backgroundColors[index % backgroundColors.length]}`}
-            />
-            <Link
-              href={`/${groupId}/tasklist/${taskList.id}`}
-              className="text-base font-medium text-white"
-            >
-              {taskList.name}
-            </Link>
-            <div className="flex items-center gap-2">
-              <div className="flex w-14 items-center gap-1 rounded-xl bg-background-primary px-2 py-1">
-                {completedItems === totalTasks ? (
-                  <IconTaskDone />
-                ) : (
-                  <PieChart width={16} height={16}>
-                    <Pie
-                      data={[{ name: 'Background', value: 100 }]}
-                      dataKey="value"
-                      cx="50%"
-                      cy="50%"
-                      innerRadius={4}
-                      outerRadius={7}
-                      startAngle={90}
-                      endAngle={-270}
-                      fill="#F8FAFC"
-                      stroke="none"
-                    />
-                    <Pie
-                      data={[
-                        { name: 'Completed', value: completionPercentage },
-                      ]}
-                      dataKey="value"
-                      cx="50%"
-                      cy="50%"
-                      innerRadius={4}
-                      outerRadius={7}
-                      startAngle={270}
-                      endAngle={270 + (completionPercentage * 360) / 100}
-                      fill="#10B981"
-                      stroke="none"
-                      cornerRadius={50}
-                    />
-                  </PieChart>
-                )}
-                <div className="text-sm font-medium text-brand-primary">
-                  {completedItems}/{totalTasks}
-                </div>
-              </div>
-              <DropdownMenu iconType="task" />
-            </div>
-          </div>
-        );
-      })}
+      {taskLists.map((taskList, index) => (
+        <TodoListItem
+          key={taskList.id}
+          taskList={taskList}
+          groupId={groupId}
+          backgroundColor={backgroundColors[index % backgroundColors.length]}
+          taskListData={data[index]}
+        />
+      ))}
     </div>
   );
 }
